turn off normalization of predicted images

+++ micro_dl/cli/image_inference.py
@@ -202,10 +202,10 @@ def run_prediction(args, gpu_ids, gpu_mem_frac):
                 )
                 file_name = os.path.join(pred_dir, im_name)
                 if args.ext == '.png' or args.ext == '.tif':
-                    # Convert to float16 and remove batch dimension
+                    # Convert to float32 and remove batch dimension
                     # im_pred = 2 ** 16 * (im_pred - im_pred.min()) /\
                     #           (im_pred.max() - im_pred.min())
-                    im_pred = np.squeeze(im_pred.astype(np.float16))
+                    im_pred = np.squeeze(im_pred.astype(np.float32))
                     cv2.imwrite(file_name, im_pred)
                 elif args.ext == '.npy':
                     np.save(file_name, im_pred, allow_pickle=True)
